Applied changes upon review

+++ src/main/java/com/example/LocalZero/dto/LikeResponse.java
@@ -6,6 +6,6 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class LikeResponse {
-    private int count;
+    private long count;
     private boolean likedByCurrentUser;
 }
